Bugfix in callback function (#119)
Due to the use of invokeCallback boolean instead of continue keyword, the piece of code to determine whether httpcallback was successful (if ret != 0) caused the callbackCounter to be increased as well. This was solved by placing the entire block of code in an if-block instead of just the args.callback and args.httpcallback.